% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_enrolled}
\alias{add_enrolled}
\title{add enrolled}
\usage{
add_enrolled(d)
}
\arguments{
\item{d}{a data frame with id, and offering or withdraw_date}
}
\value{
a data frame
}
\description{
@description \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Adds boolean field if student is enrolled
}
