% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_gpa_by}
\alias{add_gpa_by}
\title{add GPA}
\usage{
add_gpa_by(d, ...)
}
\arguments{
\item{d}{a data frame}

\item{...}{grouping variables, passed to group_by}
}
\value{
a data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Takes a data frame with \emph{grade} and other variables for grouping and
returns the a data frame with the grouping variables and \emph{gpa}.
}
