% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_grade_gpa}
\alias{add_grade_gpa}
\title{add grade gpa}
\usage{
add_grade_gpa(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Adds \emph{grade_gpa} column to data frame, based on
\emph{grade} column. Grades that count towards a GPA
are not the same as substantive grades (a key difference
being AW grades)
}
