% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_grade_helpers}
\alias{add_grade_helpers}
\title{add grade helpers}
\usage{
add_grade_helpers(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
Takes a data frame with \strong{grade} variable and returns the
same data frame with a series of grade_* variables to assist in
analysing grade data.
}
