% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_grade_npe}
\alias{add_grade_npe}
\title{add grade Non-Participating Enrolment (NPE)}
\usage{
add_grade_npe(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
Takes a data frame with \strong{grade} variable and returns the
same data frame with a \strong{grade_npe} variable, which is
true if grade is FW or FNS; and false otherwise.
}
