% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_grade_substantive}
\alias{add_grade_substantive}
\title{add grade substantive}
\usage{
add_grade_substantive(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Adds \emph{grade_substantive} column to data frame, based on
\emph{grade} column. A substantive grade is included in the denominator
when calculating progress rates.
}
