% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_grade_success}
\alias{add_grade_success}
\title{add grade success}
\usage{
add_grade_success(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Adds \emph{grade_success} column to data frame, based on
\emph{grade} column.
}
