% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_grade_zf}
\alias{add_grade_zf}
\title{add grade Zero Fail (ZF)}
\usage{
add_grade_zf(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
@description \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\details{
Takes a data frame with \strong{grade} variable and returns the
same data frame with a \strong{grade_zf} variable, which is
true if grade is FW or FNS; and false otherwise. Identical to
add_grade_npe.
}
