% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_remained_past_census}
\alias{add_remained_past_census}
\title{add remained past census}
\usage{
add_remained_past_census(d)
}
\arguments{
\item{d}{a data frame with id, session and offering}
}
\value{
a data frame
}
\description{
@description \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Adds boolean field if student is past census. This will remove
}
