% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_subject_from_offering}
\alias{add_subject_from_offering}
\title{add subject from offering}
\usage{
add_subject_from_offering(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
Takes a data frame with **offering** variable and returns the
same data frame with a **subject** variable. Requires the **offering**
to be in the format ABC123_201930_W_D, which would then return
*ABC123* as the subject.
}
