% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_year_from_offering}
\alias{add_year_from_offering}
\title{add year from offering}
\usage{
add_year_from_offering(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
Takes a data frame with **offering** variable and returns the
same data frame with a **year** variable. Requires the **offering**
to be in the format ABC123_201930_W_D which would return *2019*
}
\seealso{
Other mungers: 
\code{\link{add_year_from_session}()}
}
\concept{mungers}
