% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_year_from_session}
\alias{add_year_from_session}
\title{add year from session}
\usage{
add_year_from_session(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
Takes a data frame with **session** variable and returns the
same data frame with a **year** variable. Requires the **session**
to be in the format *201930*
}
\seealso{
Other mungers: 
\code{\link{add_year_from_offering}()}
}
\concept{mungers}
