% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affixers.R
\name{affix_academic}
\alias{affix_academic}
\title{Attaches academic summaries to a data frame}
\usage{
affix_academic(d)
}
\arguments{
\item{d}{a data frame}
}
\description{
Joins in aggregated key academic data to an existing data frame.
The grouping of the aggregation is based on the existing data frame.
}
