% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{csu_colours}
\alias{csu_colours}
\title{CSU colour palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
csu_colours
}
\description{
CSU colour palette
}
\keyword{datasets}
