% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{csu_colours_dark}
\alias{csu_colours_dark}
\title{Dark colour palette}
\format{
An object of class \code{character} of length 4.
}
\usage{
csu_colours_dark
}
\description{
Dark colour palette
}
\keyword{datasets}
