% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{csu_colours_light}
\alias{csu_colours_light}
\title{Light colour palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
csu_colours_light
}
\description{
Light colour palette
}
\keyword{datasets}
