% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{csu_colours_paired}
\alias{csu_colours_paired}
\title{CSU colour palette}
\format{
An object of class \code{character} of length 10.
}
\source{
https://about.csu.edu.au/brand/about
}
\usage{
csu_colours_paired
}
\description{
CSU colour palette
}
\keyword{datasets}
