% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_academic_by_session_summary}
\alias{fetch_academic_by_session_summary}
\title{fetch academic student session summary}
\usage{
fetch_academic_by_session_summary()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This fetches a data frame based on the enrolments and academic data that
summarises a student's result per session. Categories are \emph{pass}, \emph{fail some},
\emph{partial npe}, \emph{total npe} and \emph{withdrawn}.
}
