% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_academic_summary}
\alias{fetch_academic_summary}
\title{Fetches a summarised academic data frame}
\usage{
fetch_academic_summary()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Aggregates the academic table to one row per student per session.
}
