% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_academic_summary_by}
\alias{fetch_academic_summary_by}
\title{Fetches a summarised academic data frame}
\usage{
fetch_academic_summary_by(...)
}
\arguments{
\item{...}{variables to group output by}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

Aggregates the academic table based on grouping variables. Grouping
variables must be in the \emph{academic} table, so \emph{id}, \emph{session}, \emph{subject}
are the probably choices.
}
