% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_enrolment_counts_by}
\alias{fetch_enrolment_counts_by}
\title{fetch counts of enrolments / withdrawals}
\usage{
fetch_enrolment_counts_by(...)
}
\arguments{
\item{...}{grouping variables}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

Fetches counts of enrolled subjects, withdrawn subject and substantive grades
by grouping variables (such as \emph{id}, \emph{session}).
}
