% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_former_fails}
\alias{fetch_former_fails}
\title{Fetch former fails}
\usage{
fetch_former_fails(this_session, n_prior_sessions = 3)
}
\arguments{
\item{session}{Session code - former fails depend on the two previous sessions}
}
\value{
a data frame, one row per student
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Given a session this fetches students that have failed 50\% or more of their subjects
in the previous 1 - 3 sessions. This function collects the session results together, based
on \code{n_prior_sessions}. So if \code{n_prior_sessions == 2}, then it looks at the results in the
two prior major sessions (30, 60, 90) together.
}
