% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_student_subject_summary_from_ids}
\alias{fetch_student_subject_summary_from_ids}
\title{Gets a beefed up table of student data by subject from a list of ids}
\usage{
fetch_student_subject_summary_from_ids(
  ids,
  concerns = c("course requirement", "low activity", "non submission",
    "prior performance")
)
}
\arguments{
\item{ids}{A character vector of student ids}

\item{concerns}{A list of options from for the concern field in the flags table.}
}
\value{
a data frame, one row per student per subject per session
}
\description{
Given list of student ids this returns a data table with one row
per student per subject per session with as much demographic, retention and academic
data as possible. By default it only includes flags with the concerns:
\emph{course requirement, low activity, non submission} and \emph{prior performance}
}
\details{
You \emph{must} have the \code{retention.data} package loaded for this
to work, or the equivalent tables.
}
