% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_student_summary_from_course}
\alias{fetch_student_summary_from_course}
\title{Gets a beefed up table of student data by session based on course}
\usage{
fetch_student_summary_from_course(course_filter_string = ".")
}
\arguments{
\item{course_filter_string}{A regex expression used to filter on the course}
}
\value{
a data frame, one row per student per session
}
\description{
Given a course filtering this returns a data table with one row
per student per session with as much demographic, retention and academic
data as possible.
}
\details{
You \emph{must} have the \strong{retention.data} package loaded for this
to work, or the equivalent tables.

Some of the calculated fields are worth explaining:

* \strong{progress_rate} is the ratio of passing grades (PS, CR, DI, HD) to all grades
* \strong{pass_rate} is the ratio of passing grades to all finalised grades
* \strong{fail_rate} is the ratio of failing grades (FL, FW) to all grades
* \strong{fw_rate} is the ratio of FW (fail by non submission) to all grades
}
