% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_student_summary_from_subject}
\alias{fetch_student_summary_from_subject}
\title{Gets a beefed up table of student data by subject}
\usage{
fetch_student_summary_from_subject(
  subject_filter_string = ".",
  concerns = c("course requirement", "low activity", "non submission",
    "prior performance")
)
}
\arguments{
\item{subject_filter_string}{A regex expression used to filter on the course}

\item{concerns}{A list of options from for the concern field in the flags table.}
}
\value{
a data frame, one row per student per subject per session
}
\description{
Given a subject filtering this returns a data table with one row
per student per subject per session with as much demographic, retention and academic
data as possible. By default it only includes flags with the concerns:
\emph{course requirement, low activity, non submission} and \emph{prior performance}
}
\details{
You \emph{must} have the \code{retention.data} package loaded for this
to work, or the equivalent tables.
}
