% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_subject_demographic_summary}
\alias{fetch_subject_demographic_summary}
\title{Gets a overall demographic summary by subject}
\usage{
fetch_subject_demographic_summary(
  subject_string,
  by_session = TRUE,
  sessions = NULL,
  include_withdrawn = FALSE,
  domestic_only = TRUE
)
}
\arguments{
\item{subject_string}{a regex expression used to filter the subjects}

\item{by_session}{if TRUE then returns one row per subject per session}

\item{sessions}{a numeric vector with the sessions that the data should be based on, greedy if left as NULL}

\item{include_withdrawn}{if set to FALSE then students that withdraw are removed from the results}

\item{domestic_only}{if TRUE then only domestic enrolments are used}
}
\value{
a data frame, one row per subject (and per session if by session is TRUE)
}
\description{
Returns a summary of basic key demographics breakdown based on
a selection of subjects. Currently the key demographics selected by
this function are Low SES, Australian Indigenous, Regional and First in Family
}
