% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetct_prior_attempts}
\alias{fetct_prior_attempts}
\title{Fetch prior fails}
\usage{
fetct_prior_attempts(sub, sesh)
}
\arguments{
\item{sesh}{}
}
\description{
Given a subject and session finds current enrolments who have
previously attempted the subject
}
