% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{gpa}
\alias{gpa}
\title{GPA}
\usage{
gpa(grades)
}
\arguments{
\item{grades}{a list of grades}
}
\value{
a number, grade point average
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates GPA from a list of grades
}
