% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\docType{data}
\name{grade_classifications}
\alias{grade_classifications}
\title{grade classifications}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38 rows and 4 columns.
}
\usage{
grade_classifications
}
\description{
What each of the grade codes count towards
}
\keyword{datasets}
