% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{nice_count}
\alias{nice_count}
\title{nice count}
\usage{
nice_count(d, col)
}
\arguments{
\item{col}{Column to be counted}
}
\description{
@description \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Counts a field, then uses the janitor functions to add \% and total.
Works for a single field only
}
