% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{read_cleaned_gc_xlsx}
\alias{read_cleaned_gc_xlsx}
\title{reads grade book download with cleaned header}
\usage{
read_cleaned_gc_xlsx(f)
}
\arguments{
\item{path}{Path to the gc csv file}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Reads in a greade centre download csv, as taken from the i2 site.
The top 4 rows have the metadata: week due, value, eai & threshold mark
The labels of those columns should be in cells A1:A4
}
