% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importers.R
\name{read_gb_csv}
\alias{read_gb_csv}
\title{reads grade book download}
\usage{
read_gb_csv(path)
}
\arguments{
\item{path}{Path to the gc csv file}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Reads in a greade centre download csv, as taken from the i2 site.
File should be saved in it's default name: gc_...
}
