% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{str_c_sentence_list}
\alias{str_c_sentence_list}
\title{str_c_sentence_list}
\usage{
str_c_sentence_list(s)
}
\arguments{
\item{s}{a character vector}
}
\value{
a string
}
\description{
@description \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\details{
Takes a character vector and returns a string of
the elements in that character vector separated by
commas, with the last two elements separated with " and "
}
