% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{summarise_academic}
\alias{summarise_academic}
\title{summarise academic}
\usage{
summarise_academic(
  d,
  include_grades = FALSE,
  zf_text = "ZF",
  zf_long_text = "zero fail"
)
}
\arguments{
\item{d}{a data frame with grades in it, grouped by relevant}

\item{include_grades}{true / false: do you want to include a list of the grades as well?}

\item{zf_text}{The text for ZF / NPE / FNS grades}

\item{zf_long_text}{The longer text for 'zero fail' or 'non-participating enrolment' or 'FNS'}
}
\value{
a summarised data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates a summary of grades.
}
