% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_grade_fail}
\alias{add_grade_fail}
\title{add grade fail}
\usage{
add_grade_fail(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
Takes a data frame with \strong{grade} variable and returns the
same data frame with a \strong{grade_fail} variable, which is
true if grade is FW, FL; and false otherwise.
}
