% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_grade_finalised}
\alias{add_grade_finalised}
\title{add grade finalised}
\usage{
add_grade_finalised(d)
}
\arguments{
\item{d}{a data frame}
}
\value{
a data frame
}
\description{
Takes a data frame with \strong{grade} variable and returns the
same data frame with a \strong{grade_finalised} variable, which is
true if grade is FW, FL, PS, CR, DI or HD; and false otherwise.
}
