% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungers.R
\name{add_subject_context}
\alias{add_subject_context}
\title{add subject context}
\usage{
add_subject_context(d)
}
\arguments{
\item{d}{data frame, as specified above}
}
\description{
@description \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\details{
Requires a data frame with the \emph{subject} (six character code),
\emph{n_remained} (if this is NA the subject wasn't in HEPPP last year), \emph{genuine_pass_rate},
\emph{n_passed},  \emph{same_trigger} (boolean) and also
three fields describing when the subject runs, \emph{s1}, \emph{s2} and \emph{s3}.
The \emph{sn} fields should be set to \strong{internal}, \strong{distance} or \strong{both}.
If they are all blank (NA) then it is deemed that it is unknown when the subject is run next.
}
\examples{
d <- tibble::tribble(
   ~subject, ~genuine_pass_rate, ~n_remained, ~n_passed, ~same_trigger, ~s1, ~s2, ~s3,
   "MEH111", NA_real_, NA, NA, NA, "both", "internal", "internal",
   "NUP200", 0.88, 18, 3, TRUE, NA_character_, "both", NA_character_,
   "LOL101", 0.71, 15, 0, FALSE, "distance", "distance", "distance"
)
d |> add_subject_context()

}
