% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affixers.R
\name{affix_demographics}
\alias{affix_demographics}
\title{Attaches demographic summaries to a data frame}
\usage{
affix_demographics(d)
}
\arguments{
\item{d}{a data frame}
}
\description{
Joins in aggregated key demographic data to an existing data frame.
The grouping of the aggregation is based on the existing data frame.
As it is aggregated the categorical variables are reduced to a count and
a relevant percentage of the key categorical value of interest.
}
