% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_prior_attempts}
\alias{fetch_prior_attempts}
\title{Fetch prior attempts}
\usage{
fetch_prior_attempts(enr, aca, sub, sesh)
}
\arguments{
\item{enr}{data frame of enrolments}

\item{aca}{data frame of academic results}

\item{sub}{subject}

\item{sesh}{session}
}
\description{
Given a subject and session finds current enrolments who have
previously attempted the subject
}
