% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_student_session_summary_from_ids}
\alias{fetch_student_session_summary_from_ids}
\title{Gets a beefed up table of student data by session based on a list of student ids}
\usage{
fetch_student_session_summary_from_ids(
  ids,
  concerns = c("course requirement", "low activity", "non submission",
    "prior performance")
)
}
\arguments{
\item{ids}{A character vector of student ids}

\item{concerns}{A list of options from for the concern field in the flags table.}
}
\value{
a data frame, one row per student per session
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Given a list of student ids this returns a data table with one row
per student per session with as much demographic, retention and academic
data as possible. By default it only includes flags with the concerns:
\emph{course requirement, low activity, non submission} and \emph{prior performance}

You \emph{must} have the \code{retention.data} package loaded for this
to work, or the equivalent tables.

Some of the calculated fields are worth explaining:
\itemize{
\item \strong{progress_rate} is the ratio of passing grades (PS, CR, DI, HD) to all grades
\item \strong{pass_rate} is the ratio of passing grades to all finalised grades
\item \strong{fail_rate} is the ratio of failing grades (FL, FW) to all grades
\item \strong{fw_rate} is the ratio of FW (fail by non submission) to all grades
}
}
